// frontenditsol-master/src/app/model/nhanvien.ts
import { PhongBan } from './phongban';

export class NhanVien{
             id: number;
            tenNhanVien: string;
            ngaySinh: string;
            gioiTinh: boolean;
            diaChi: string;
            soCmnnd: string;
            dienThoai: string;
            chuVu: string;
            email: string;
            img: string;
            action: boolean;
            phongBan: PhongBan;
            phongbanid:number;
            ngayBatDau : string;
            ngayKetThuc : string;
            isSelected;

}